feat: Add Appointment feature with booking and payment (UI)
- Implemented `AppointmentScreen` with stepper navigation for Date & Time, Payment, and Summary steps.
- Added `DateAndTimeContent` for selecting appointment date, time, and type.
- Integrated `PaymentAppointmentContent` for payment method selection.
- Created `SummaryAppointmentContent` to display booking and payment summaries.
- Developed reusable widgets like `DatePickerAppointmentScreen`, `TimeAppointmentScreen`, and `AppointmentTypeSummary`.
- Integrated `AppointmentCubit` for state management of appointment type, payment method, date, and time.
- Enhanced UI with custom styles and responsive layouts.

+++ lib/features/home/ui/widgets/speciality_list/categories_doctor_listView.dart
@@ -30,8 +30,9 @@ class _CategoriesDoctorListViewState extends State<CategoriesDoctorListView> {
                 setState(() {
                   selectedSpecializationIndex = index;
                 });
-                context.read<HomeCubit>().getListDoctors(
-                    id: widget.homeModels.data[index].id!.toInt());
+                context
+                    .read<HomeCubit>()
+                    .getListDoctors(id: widget.homeModels.data[index].id!);
               },
               child: CategoriesDoctorItems(
                 specialization: widget.homeModels.data[index],
